requirements.txt
anthropic>=0.40.0
google-generativeai>=0.8.0
typer>=0.12.0
rich>=13.7.0
ffmpeg-python>=0.2.0
moviepy==1.0.3
python-dotenv>=1.0.0
openai-whisper>=20231117
numpy>=1.24.0
